refactor : Product와 ProductImage 병합
