+++ src/main/java/com/be/rebook/domain/product/repository/ProductImageRepository.java
@@ -1,6 +1,6 @@
-package com.be.rebook.domain.productImage.repository;
+package com.be.rebook.domain.product.repository;
 
-import com.be.rebook.domain.productImage.domain.ProductImage;
+import com.be.rebook.domain.product.entity.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
