+++ src/main/java/com/be/rebook/domain/product/service/ProductService.java
@@ -7,8 +7,8 @@ import com.be.rebook.domain.members.repository.MembersRepository;
 import com.be.rebook.domain.product.entity.Product;
 import com.be.rebook.domain.product.dto.ProductRequestDTO;
 import com.be.rebook.domain.product.repository.ProductRepository;
-import com.be.rebook.domain.productImage.domain.ProductImage;
-import com.be.rebook.domain.productImage.repository.ProductImageRepository;
+import com.be.rebook.domain.product.entity.ProductImage;
+import com.be.rebook.domain.product.repository.ProductImageRepository;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
